Module Loan : Action rejeter et accepter requêtes

+++ angular-frontend/src/app/loan/components/loan-requests/loan-requests.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Observable } from 'rxjs';
+import { DynamicFormQuestion, HiddenQuestion, DisplayQuestion } from 'src/app/shared/components/dynamic-form-question/dynamic-form-question.component';
 import { CustomColumn } from 'src/app/shared/components/table/custom-column';
 import { TableComponent } from 'src/app/shared/components/table/table.component';
 import { CreateEditDeleteDialogService } from 'src/app/shared/services/create-edit-delete-dialog.service';
@@ -16,8 +17,17 @@ export class LoanRequestsComponent implements OnInit {
 
   @ViewChild('table') table: TableComponent;
 
-  acceptAction = this.createEditDeleteDialogService.buildBlankAction();
-  rejectAction = this.createEditDeleteDialogService.buildBlankAction();
+  questions: DynamicFormQuestion[] = [
+    new DisplayQuestion({ key: 'applicant', label: 'TABLE.LOAN-REQUEST-USER' }),
+    new DisplayQuestion({ key: 'resource_name', label: 'TABLE.LOAN-REQUEST-RESOURCE' }),
+    new DisplayQuestion({ key: 'request_type_name', label: 'TABLE.LOAN-REQUEST-TYPE', translated: true }),
+    new DisplayQuestion({ key: 'start_date', label: 'TABLE.LOAN-REQUEST-START-DATE', type: 'date' }),
+    new DisplayQuestion({ key: 'end_date', label: 'TABLE.LOAN-REQUEST-END-DATE', type: 'date' }),
+    new HiddenQuestion({ key: 'id' })
+  ];
+
+  acceptAction = this.createEditDeleteDialogService.buildDisplayAction(this.questions, 'primary', 'TABLE.LOAN-REQUEST-ACCEPT', 'LOAN.ACCEPT-DIALOG-TITLE');
+  rejectAction = this.createEditDeleteDialogService.buildDisplayAction(this.questions, 'warn', 'TABLE.LOAN-REQUEST-REJECT', 'LOAN.REJECT-DIALOG-TITLE');
 
   columns: CustomColumn[] = [
     {name: 'TABLE.LOAN-REQUEST-USER', property: 'applicant'},
@@ -45,11 +55,15 @@ export class LoanRequestsComponent implements OnInit {
     this.crossComponentService.title = 'SIDENAV.LOAN.MANAGE-REQUESTS';
 
     this.acceptAction.subject.subscribe((data) => {
-      this.loanService.acceptLoanRequest(data).subscribe(this.loanService.api.deleteTableUpdate(this.table)(data));
+      if (data !== null) {
+        this.loanService.acceptLoanRequest(data).subscribe(this.loanService.api.deleteTableUpdate(this.table));
+      }
     });
 
     this.rejectAction.subject.subscribe((data) => {
-      this.loanService.rejectLoanRequest(data).subscribe(this.loanService.api.deleteTableUpdate(this.table));
+      if (data !== null) {
+        this.loanService.rejectLoanRequest(data).subscribe(this.loanService.api.deleteTableUpdate(this.table));
+      }
     });
   }
 }
